Split webhooks types
