+++ packages/types/src/discord.ts
@@ -2,7 +2,8 @@ import type { DiscordApplication } from './discord/applications.js'
 import type { ChannelTypes, DiscordChannel, DiscordOverwrite, DiscordThreadMember, SortOrderTypes } from './discord/channels.js'
 import type { DiscordGuild, DiscordMemberWithUser } from './discord/guilds.js'
 import type { DiscordUser } from './discord/users.js'
-import type { RoleFlags, TeamMembershipStates, WebhookTypes } from './shared.js'
+import type { DiscordIncomingWebhook } from './discord/webhooks.js'
+import type { RoleFlags, TeamMembershipStates } from './shared.js'
 
 /** https://discord.com/developers/docs/topics/oauth2#shared-resources-oauth2-scopes */
 export enum OAuth2Scope {
@@ -221,65 +222,6 @@ export interface DiscordTeamMember {
   user: Partial<DiscordUser> & Pick<DiscordUser, 'avatar' | 'discriminator' | 'id' | 'username' | 'global_name'>
   /** Role of the team member */
   role: DiscordTeamMemberRole
-}
-
-/** https://discord.com/developers/docs/resources/webhook#webhook-object-webhook-structure */
-export type DiscordWebhook = DiscordIncomingWebhook | DiscordApplicationWebhook
-
-export interface DiscordIncomingWebhook {
-  /** The type of the webhook */
-  type: WebhookTypes
-  /** The secure token of the webhook (returned for Incoming Webhooks) */
-  token?: string
-  /** The url used for executing the webhook (returned by the webhooks OAuth2 flow) */
-  url?: string
-
-  /** The id of the webhook */
-  id: string
-  /** The guild id this webhook is for */
-  guild_id?: string
-  /** The channel id this webhook is for */
-  channel_id: string
-  /** The user this webhook was created by (not returned when getting a webhook with its token) */
-  user?: DiscordUser
-  /** The default name of the webhook */
-  name: string | null
-  /** The default user avatar hash of the webhook */
-  avatar: string | null
-  /** The bot/OAuth2 application that created this webhook */
-  application_id: string | null
-  /** The guild of the channel that this webhook is following (returned for Channel Follower Webhooks) */
-  source_guild?: Partial<DiscordGuild>
-  /** The channel that this webhook is following (returned for Channel Follower Webhooks) */
-  source_channel?: Partial<DiscordChannel>
-}
-
-export interface DiscordApplicationWebhook {
-  /** The type of the webhook */
-  type: WebhookTypes.Application
-  /** The secure token of the webhook (returned for Incoming Webhooks) */
-  token?: string
-  /** The url used for executing the webhook (returned by the webhooks OAuth2 flow) */
-  url?: string
-
-  /** The id of the webhook */
-  id: string
-  /** The guild id this webhook is for */
-  guild_id?: string | null
-  /** The channel id this webhook is for */
-  channel_id?: string | null
-  /** The user this webhook was created by (not returned when getting a webhook with its token) */
-  user?: DiscordUser
-  /** The default name of the webhook */
-  name: string | null
-  /** The default user avatar hash of the webhook */
-  avatar: string | null
-  /** The bot/OAuth2 application that created this webhook */
-  application_id: string | null
-  /** The guild of the channel that this webhook is following (returned for Channel Follower Webhooks), field will be absent if the webhook creator has since lost access to the guild where the followed channel resides */
-  source_guild?: Partial<DiscordGuild>
-  /** The channel that this webhook is following (returned for Channel Follower Webhooks), field will be absent if the webhook creator has since lost access to the guild where the followed channel resides */
-  source_channel?: Partial<DiscordChannel>
 }
 
 /** https://discord.com/developers/docs/topics/permissions#role-object-role-structure */
