+++ packages/types/src/discord/auditLogs.ts
@@ -1,6 +1,6 @@
 /** Types for: https://discord.com/developers/docs/resources/audit-log */
 
-import type { DiscordRole, DiscordWebhook } from '../discord.js'
+import type { DiscordRole } from '../discord.js'
 import type { DiscordAutoModerationRule } from './autoModeration.js'
 import type { DiscordChannel, DiscordOverwrite, DiscordThreadMetadata } from './channels.js'
 import type { DiscordEmoji } from './emojis.js'
@@ -11,6 +11,7 @@ import type { DiscordInvite, DiscordInviteMetadata } from './invites.js'
 import type { DiscordStageInstance } from './stageInstances.js'
 import type { DiscordSticker } from './stickers.js'
 import type { DiscordUser } from './users.js'
+import type { DiscordWebhook } from './webhooks.js'
 
 /** https://discord.com/developers/docs/resources/audit-log#audit-log-object-audit-log-structure */
 export interface DiscordAuditLog {
