+++ packages/types/src/index.ts
@@ -21,6 +21,7 @@ export * from './discord/stickers.js'
 export * from './discord/subscriptions.js'
 export * from './discord/users.js'
 export * from './discord/voice.js'
+export * from './discord/webhooks.js'
 export * from './discord/webhookEvents.js'
 export * from './discordeno.js'
 export * from './shared.js'
