+++ packages/types/src/shared.ts
@@ -15,16 +15,6 @@ export enum TeamMembershipStates {
   Accepted,
 }
 
-/** https://discord.com/developers/docs/resources/webhook#webhook-object-webhook-types */
-export enum WebhookTypes {
-  /** Incoming Webhooks can post messages to channels with a generated token */
-  Incoming = 1,
-  /** Channel Follower Webhooks are internal webhooks used with Channel Following to post new messages into channels */
-  ChannelFollower,
-  /** Application webhooks are webhooks used with Interactions */
-  Application,
-}
-
 /** https://discord.com/developers/docs/topics/permissions#permissions-bitwise-permission-flags */
 export const BitwisePermissionFlags = {
   /** Allows creation of instant invites */
